feat: Replaced 'Next' components to default
